+++ app/controllers/external_controller.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+# Displays public pages.
+class ExternalController < ApplicationController
+  # # GET /privacy-policy
+  # def privacy_policy
+  # end
+end