+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## 19.0.0
 
+### Enhancements
+- **General Changes**
+  - Add updated Privacy Policy
+
 ## 18.0.0 (April 17, 2018)
 
 ### Enhancements